Host implementation of JPEG parser
The goal is scaling down the JPEG by transcoding.

The code is written from scratch, although I am using libjpeg as a
reference. The idea is to get the code to work on the CPU, and then
port it to the DPU. Therefore, I am writing code that will be easy
to implement on the DPU.

Signed-off-by: Joel Nider <joel@nider.org>