feat: separated JpegDecompressor to 2 structs. Partially done parallel huffman decoding
JpegDecompressor is now split into a global struct which contains all
shared variables between tasklets such as quantization tables and
huffman tables, and the other part is a per tasklet struct which holds
information about the current position in the file and the bit buffer.
This is essential for parallel decoding. For parallel decoding, Huffman
coded bitstream is evenly split amongst all tasklets to decode, and all
tasklets are currently only decoding their own assigned part.