Reorganize the requirements for the metrics
Pillow is a prerequisite for scipi, which means pip will try to install
it first. My system had an older version of pillow (7.0.0) already
installed that satisfied the requirement but it really didn't. Scifi
requires pillow>=8.3.4 so it couldn't be installed. I'm not sure why
pip didn't fix this automatically, but I'm not a pip expert (and don't
aim to be) so I'm not going hunting for a lot of weird flags to force
it to have normal behaviour out-of-the-box. Instead, I updated the
requirements.txt to (1) have pillow first, and (2) specify a minimal
version. Of course, this is likely to break as soon as we move to a new
system that has an updated scifi package, and needs a newer version
(which seems to me package version management and prerequisites are
part of the reason that pip exists in the first place?)

Well, it seems to work a bit better but may need more tweaking in the
future.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>