Add script for image classification
As a potential test of JPEG suitability, this DNN uses a pretrained
model to classify a set of images. The nice feature is that the input
images can be in a range of formats and opencv takes care of decoding
them for us. So it is quite easy to run the same workload on two sets
of images: JPEG encoded originals, and BMP versions decoded by PIM.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>