Ensure all image data is copied to host
The size of the raw image data is not calculated properly. The data
is often truncated while being copied from the DPUs back to the host.
This is a temporary fix to the problem. Ideally, the size of the data
will be calculated by the DPU, and returned as a variable.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>