Enable operating multiple ranks asynchronously
As files are prepared (i.e. read from the file system) they are
copied into memory and pushed to the DPUs, trying to fill a rank.
As soon as a rank is full, it is launched asynchronously, and the
next batch is prepared. As long as there are free ranks in hardware
more files will be prepared and launched. When there are no more
free ranks but there are waiting files, the program will poll for
completion of any running rank.

Tested on the simulator with:
./host-1 -d data/imagenet/*.JPEG

Signed-off-by: Joel Nider <joel@ece.ubc.ca>