Check upper bound on buffers
The maximum size of a decoded image is limited by the amount of MRAM.
Currently, the buffer is set to 32MB-512B and is not checked
anywhere. This causes memory overflows and DPU faults when trying to
write beyond the limits of the variable (overwriting other variables)
and sometimes beyond the physical MRAM limits. It wastes time for
the user, who thinks the image is being decompressed, when the
program could know that it is not possible to complete.

I raised this to a full 32MB (not sure why it was slightly less to
begin with) and put it in a macro.
The DPU now calculates the maximum size of the decompressed image,
based on the MCU size, image dimensions and color depth (number of
planes). This number is compared to the size of the allocated space
for the decompressed image, and immediately returns if the image
will not fit.
Since DPUs are launched at rank granularity, it is possible that
some DPUs will succeed to decode their images, while others might
fail. In this case, only the ones that succeed will be copied back
and written to BMP files. Therefore, I included an additional check
(set the image size to 0) for any DPUs that failed decoding.

At the same time, I have applied some other minor optimizations:
1) Reduce file size variable to 32-bits - no point in using a 64-bit
variable on the DPU for this, since we will never have an image
larger than 4GB anyway.
2) Use a 'flags' field for binary options, such as "horizontal flip".
We can cram in a lot more flags this way, reducing the amount of
data that has to be written to each DPU at launch.