Add scalability test
To measure how the DPUs perform as more images are added, add the -S
option. The -m option can then be used to set the number of files
that will be decoded. If not enough files are provided on the
command line, the last file will be reused to fill in the remaining
slots. Thus, the maximum number of files is always tested, as long as
at least one file name is provided.

Build with STATS=1 to print out the total decoding time.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>