Take invalid DPUs into account
Some ranks have bad DPUs. That means we cannot rely on the runtime
calculation of how many DPUs are in each rank. Rather, we must
allocate space for the maximum number of DPUs, even if they are not
all used. Otherwise, we get errors when trying to read from a
non-existent buffer.

This is based on Niloo's change:
https://github.com/UBC-ECE-Sasha/PIM-JPEG/commit/93459a3c1589e8513f2cf603937d5ca4c19842b5